Article findAll

+++ project-01-rest-jdbctemplate/src/main/java/com/example/project01restjdbctemplate/model/Article.java
@@ -1,12 +1,11 @@
 package com.example.project01restjdbctemplate.model;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 public class Article {
 	
 	
-	private long id;
+	private int id;
 	private String title;
 	private String body;
 	private LocalDateTime createdAt;
